+++ tests/data/utils/invalid/deploy_notjson
@@ -1,0 +1,1 @@
+This is not a json file
