+++ tests/data/utils/invalid/develop.json
@@ -1,0 +1,76 @@
+[
+  "/home/pombredanne/bin/tracecode/tst/test_data/analy_patchelf/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/analy_patchelf_settings/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/analy_patchelf_settings2/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/as_ops/trace/t.2799712",
+  "/home/pombredanne/bin/tracecode/tst/test_data/bug1/t.2461545",
+  "/home/pombredanne/bin/tracecode/tst/test_data/can_parse_dup_call/trace/bash-build01.70793",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_cups_settings/expected",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_cups_settings/sources",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_cups_settings/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_patchelf_settings/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_strace_settings/expected",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_strace_settings/sources",
+  "/home/pombredanne/bin/tracecode/tst/test_data/d2d_strace_settings/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/file_graph_analysis_patchelf/expected_ints",
+  "/home/pombredanne/bin/tracecode/tst/test_data/file_graph_analysis_patchelf/expected_srcs",
+  "/home/pombredanne/bin/tracecode/tst/test_data/file_graph_analysis_patchelf/expected_tgts",
+  "/home/pombredanne/bin/tracecode/tst/test_data/file_graph_analysis_patchelf/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/filtering/t/trace.70676",
+  "/home/pombredanne/bin/tracecode/tst/test_data/grep/trace/t.2799712",
+  "/home/pombredanne/bin/tracecode/tst/test_data/grep2/trace/t.2799712",
+  "/home/pombredanne/bin/tracecode/tst/test_data/grep2/trace/t.2799713",
+  "/home/pombredanne/bin/tracecode/tst/test_data/grep3/trace/t.2799712",
+  "/home/pombredanne/bin/tracecode/tst/test_data/grep3/trace/t.2799713",
+  "/home/pombredanne/bin/tracecode/tst/test_data/grep3/trace/t.2799714",
+  "/home/pombredanne/bin/tracecode/tst/test_data/interleaving/trace/cp-trace.1",
+  "/home/pombredanne/bin/tracecode/tst/test_data/interleaving/trace/cp-trace.2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/interleaving/trace/t.15480",
+  "/home/pombredanne/bin/tracecode/tst/test_data/lists/trace/t.2799712",
+  "/home/pombredanne/bin/tracecode/tst/test_data/lists_patchelf/expected",
+  "/home/pombredanne/bin/tracecode/tst/test_data/lists_patchelf/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/parse_does_not_hang/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/parse_entry/basic.trace",
+  "/home/pombredanne/bin/tracecode/tst/test_data/parse_error/trace.1",
+  "/home/pombredanne/bin/tracecode/tst/test_data/patchelf_command/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/patchelf_command2/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/patchelf_command2/ignored_reads.lst",
+  "/home/pombredanne/bin/tracecode/tst/test_data/patchelf_command2/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_cups/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_cups/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_grep2/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_grep2/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_grep3/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_grep3/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_patchelf/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_patchelf/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_cleaning/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_cleaning/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_cleaning_dir/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_cleaning_dir/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_full/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_full/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_saved/expected_pformats.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/pformat_strace_saved/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph/expected.dot",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph/trace/t.2799712",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_bash/expected.dot",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_bash/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_cups/expected.dot",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_cups/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_patchelf/expected.dot",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_patchelf/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_patchelf2/expected.dot",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_patchelf2/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_strace/expected.dot",
+  "/home/pombredanne/bin/tracecode/tst/test_data/render_graph_strace/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/stats_patchelf/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/strace_mini/trace/strace.2453402",
+  "/home/pombredanne/bin/tracecode/tst/test_data/strace_mini/trace/strace.2453403",
+  "/home/pombredanne/bin/tracecode/tst/test_data/strace_mini2/trace/strace.2453402",
+  "/home/pombredanne/bin/tracecode/tst/test_data/strace_mini2/trace/strace.2453403",
+  "/home/pombredanne/bin/tracecode/tst/test_data/strace2/trace/strace.2453402",
+  "/home/pombredanne/bin/tracecode/tst/test_data/strace2/trace/strace2.2453403",
+  "/home/pombredanne/bin/tracecode/tst/test_data/validate/trace-archive.tar.bz2",
+  "/home/pombredanne/bin/tracecode/tst/test_data/validate_out_of_order/trace-archive.tar.bz2"
+]
